% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businesslogic.R
\name{BusinessLogic}
\alias{BusinessLogic}
\title{R6 Class for the business logic of the app}
\description{
It does the data cleaning and generates text and chart for the app.
}
\examples{

## ------------------------------------------------
## Method `BusinessLogic$new`
## ------------------------------------------------

business_logic <- BusinessLogic$new()

business_logic$generate_results("9E", 1, "arr_delay", 30)

business_logic$results$title

business_logic$results$text

business_logic$results$chart
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{months}}{Month numbers and their names.}

\item{\code{metrics}}{Available metrics.}

\item{\code{data}}{Will store the data..}

\item{\code{unique_carriers}}{Available carriers.}

\item{\code{results}}{Where results are stored.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BusinessLogic$new()}}
\item \href{#method-generate_results}{\code{BusinessLogic$generate_results()}}
\item \href{#method-clone}{\code{BusinessLogic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new business logic object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessLogic$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new `BusinessLogic` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{business_logic <- BusinessLogic$new()

business_logic$generate_results("9E", 1, "arr_delay", 30)

business_logic$results$title

business_logic$results$text

business_logic$results$chart
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_results"></a>}}
\if{latex}{\out{\hypertarget{method-generate_results}{}}}
\subsection{Method \code{generate_results()}}{
Generate results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessLogic$generate_results(
  carrier_filter,
  month_filter,
  metric_filter,
  threshold
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{carrier_filter}}{Carrier for report.}

\item{\code{month_filter}}{Month for report.}

\item{\code{metric_filter}}{Metric for report.}

\item{\code{threshold}}{Threshold for report.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessLogic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
